requirements.txt
numpy==1.6.2
Pillow==1.1
PyYAML==3.11
pytz==2015.7
pyparsing==2.0.2
six==1.9.0
scipy==0.10.1
matplotlib==1.3.1
python-dateutil==2.2
nose
mock
